+++ JamfUploaderProcessors/JamfComputerProfileUploader.py
@@ -1,15 +1,26 @@
 #!/usr/local/autopkg/python
 
 """
-JamfComputerProfileUploader processor for uploading a computer configuration profile
-to Jamf Pro using AutoPkg
-    by G Pugh
+Copyright 2023 Graham Pugh
 
-Note that all functions are in JamfUploaderLib/JamfAccountUploaderBase.py
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+NOTES:
+All functions are in JamfUploaderLib/JamfComputerProfileUploaderBase.py
 """
 
-import os
-import sys
+import os.path
+import sys.path
 
 # to use a base module in AutoPkg we need to add this path to the sys.path.
 # this violates flake8 E402 (PEP8 imports) but is unavoidable, so the following
@@ -28,6 +39,7 @@ class JamfComputerProfileUploader(JamfComputerProfileUploaderBase):
         "A processor for AutoPkg that will upload a computer configuration "
         "profile to a Jamf Cloud or on-prem server."
     )
+
     input_variables = {
         "JSS_URL": {
             "required": True,
@@ -124,8 +136,6 @@ class JamfComputerProfileUploader(JamfComputerProfileUploaderBase):
             "description": "Description of interesting results.",
         },
     }
-
-    description = __doc__
 
     def main(self):
         """Run the execute function"""
