+++ JamfUploaderProcessors/JamfExtensionAttributeUploader.py
@@ -1,9 +1,22 @@
 #!/usr/local/autopkg/python
 
 """
-JamfExtensionAttributeUploader processor for uploading an extension attribute
-to Jamf Pro using AutoPkg
-    by G Pugh
+Copyright 2023 Graham Pugh
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+NOTES:
+All functions are in JamfUploaderLib/JamfExtensionAttributeUploaderBase.py
 """
 
 import os
@@ -26,6 +39,7 @@ class JamfExtensionAttributeUploader(JamfExtensionAttributeUploaderBase):
         "A processor for AutoPkg that will upload an Extension Attribute item to a "
         "Jamf Cloud or on-prem server."
     )
+
     input_variables = {
         "JSS_URL": {
             "required": True,
@@ -91,8 +105,6 @@ class JamfExtensionAttributeUploader(JamfExtensionAttributeUploaderBase):
             "description": "Description of interesting results.",
         },
     }
-
-    description = __doc__
 
     def main(self):
         """Run the execute function"""
