+++ JamfUploaderProcessors/JamfPackageUploader.py
@@ -1,17 +1,29 @@
 #!/usr/local/autopkg/python
 
 """
-JamfPackageUploader processor for AutoPkg
-    by G Pugh
+Copyright 2023 Graham Pugh
 
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+NOTES:
 Developed from an idea posted at
     https://www.jamf.com/jamf-nation/discussions/27869#responseChild166021
 
-Note that all functions are in JamfUploaderLib/JamfPackageUploaderBase.py
+All functions are in JamfUploaderLib/JamfPackageUploaderBase.py
 """
 
-import os
-import sys
+import os.path
+import sys.path
 
 # to use a base module in AutoPkg we need to add this path to the sys.path.
 # this violates flake8 E402 (PEP8 imports) but is unavoidable, so the following
@@ -241,8 +253,6 @@ class JamfPackageUploader(JamfPackageUploaderBase):
             "description": "Description of interesting results.",
         },
     }
-
-    description = __doc__
 
     def main(self):
         """Run the execute function"""
