+++ JamfUploaderProcessors/JamfPolicyUploader.py
@@ -1,12 +1,26 @@
 #!/usr/local/autopkg/python
 
 """
-JamfPolicyUploader processor for uploading policies to Jamf Pro using AutoPkg
-    by G Pugh
+Copyright 2023 Graham Pugh
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+NOTES:
+All functions are in JamfUploaderLib/JamfPolicyUploaderBase.py
 """
 
-import os
-import sys
+import os.path
+import sys.path
 
 # to use a base module in AutoPkg we need to add this path to the sys.path.
 # this violates flake8 E402 (PEP8 imports) but is unavoidable, so the following
@@ -26,6 +40,7 @@ class JamfPolicyUploader(JamfPolicyUploaderBase):
         "on-prem server. Optionally, an icon can be uploaded and associated "
         "with the policy."
     )
+
     input_variables = {
         "JSS_URL": {
             "required": True,
@@ -105,8 +120,6 @@ class JamfPolicyUploader(JamfPolicyUploaderBase):
             "description": "Jamf object ID of the newly created or modified policy.",
         },
     }
-
-    description = __doc__
 
     def main(self):
         """Run the execute function"""
