+++ JamfUploaderProcessors/JamfUploaderLib/JamfCategoryUploaderBase.py
@@ -1,12 +1,23 @@
 #!/usr/local/autopkg/python
 
 """
-Base classes for uploading a category to Jamf Pro
-    by G Pugh
+Copyright 2023 Graham Pugh
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
 """
 
 import os.path
-import sys
+import sys.path
 
 from time import sleep
