+++ JamfUploaderProcessors/JamfUploaderLib/JamfIconUploaderBase.py
@@ -1,8 +1,19 @@
 #!/usr/local/autopkg/python
 
 """
-Base classes for uploading an icon to Jamf Pro
-    by G Pugh
+Copyright 2023 Graham Pugh
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
 """
 
 import os.path
@@ -23,7 +34,7 @@ from JamfUploaderBase import JamfUploaderBase  # noqa: E402
 
 
 class JamfIconUploaderBase(JamfUploaderBase):
-    """Class for functions used to upload an extension attribute to Jamf"""
+    """Class for functions used to upload an icon to Jamf"""
 
     def get_icon(self, icon_uri):
         """download an icon file"""
