+++ JamfUploaderProcessors/JamfUploaderLib/JamfMobileDeviceProfileUploaderBase.py
@@ -23,7 +23,10 @@ import subprocess
 import uuid
 
 from time import sleep
-from autopkglib import ProcessorError  # pylint: disable=import-error  # noqa: E402
+
+from autopkglib import (
+    ProcessorError,
+)  # pylint: disable=import-error
 
 # to use a base module in AutoPkg we need to add this path to the sys.path.
 # this violates flake8 E402 (PEP8 imports) but is unavoidable, so the following
@@ -33,13 +36,13 @@ sys.path.insert(0, os.path.dirname(__file__))
 from JamfUploaderBase import JamfUploaderBase  # noqa: E402
 
 
-class JamfComputerProfileUploaderBase(JamfUploaderBase):
-    """Class for functions used to upload a computer configuration profile to Jamf"""
+class JamfMobileDeviceProfileUploaderBase(JamfUploaderBase):
+    """Class for functions used to upload a mobile device configuration profile to Jamf"""
 
     def get_existing_uuid_and_identifier(self, jamf_url, obj_id, token):
         """return the existing UUID to ensure we don't change it"""
         # first grab the payload from the xml object
-        obj_type = "os_x_configuration_profile"
+        obj_type = "configuration_profile"
         existing_plist = self.get_api_obj_value_from_id(
             jamf_url,
             obj_type,
@@ -81,72 +84,6 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
         #     plistlib.dump(mobileconfig_contents, file)
         return mobileconfig_contents
 
-    def make_mobileconfig_from_payload(
-        self,
-        payload_path,
-        payload_identifier,
-        mobileconfig_name,
-        organization,
-        description,
-        mobileconfig_uuid,
-    ):
-        """create a mobileconfig file using a payload file"""
-        # import plist as text and replace any substitutable keys
-        with open(payload_path, "rb") as file:
-            payload_text = file.read()
-        # substitute user-assignable keys (requires decode to string)
-        payload_text = self.substitute_assignable_keys(
-            (payload_text.decode()), xml_escape=True
-        )
-        # now convert to data (requires encode back to bytes...)
-        mcx_preferences = plistlib.loads(str.encode(payload_text))
-
-        self.output("Preferences contents:", verbose_level=2)
-        self.output(mcx_preferences, verbose_level=2)
-
-        # generate a random UUID for the payload
-        payload_uuid = str(uuid.uuid4())
-
-        # add the other keys required in the payload
-        payload_contents = {
-            "PayloadDisplayName": "Custom Settings",
-            "PayloadIdentifier": payload_uuid,
-            "PayloadOrganization": "JAMF Software",
-            "PayloadType": "com.apple.ManagedClient.preferences",
-            "PayloadUUID": payload_uuid,
-            "PayloadVersion": 1,
-            "PayloadContent": {
-                payload_identifier: {
-                    "Forced": [{"mcx_preference_settings": mcx_preferences}]
-                }
-            },
-        }
-
-        self.output("Payload contents:", verbose_level=2)
-        self.output(payload_contents, verbose_level=2)
-
-        # now write the mobileconfig file
-        mobileconfig_data = {
-            "PayloadDescription": description,
-            "PayloadDisplayName": mobileconfig_name,
-            "PayloadOrganization": organization,
-            "PayloadRemovalDisallowed": True,
-            "PayloadScope": "System",
-            "PayloadType": "Configuration",
-            "PayloadVersion": 1,
-            "PayloadIdentifier": mobileconfig_uuid,
-            "PayloadUUID": mobileconfig_uuid,
-            "PayloadContent": [payload_contents],
-        }
-
-        self.output("Converting config data to plist")
-        mobileconfig_plist = plistlib.dumps(mobileconfig_data)
-
-        self.output("Mobileconfig contents:", verbose_level=2)
-        self.output(mobileconfig_plist.decode("UTF-8"), verbose_level=2)
-
-        return mobileconfig_plist
-
     def unsign_signed_mobileconfig(self, mobileconfig_plist):
         """checks if profile is signed. This is necessary because Jamf cannot
         upload a signed profile, so we either need to unsign it, or bail"""
@@ -178,11 +115,10 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
         description,
         category,
         mobileconfig_plist,
-        computergroup_name,
+        devicegroup_name,
         template_contents,
         profile_uuid,
         token,
-        retain_scope=False,
         obj_id=0,
     ):
         """Update Configuration Profile metadata."""
@@ -199,7 +135,7 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
             "description": description,
             "category": category,
             "payload": mobileconfig,
-            "computergroup_name": computergroup_name,
+            "devicegroup_name": devicegroup_name,
             "uuid": f"com.github.grahampugh.jamf-upload.{profile_uuid}",
         }
 
@@ -219,21 +155,12 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
         self.output("Configuration Profile to be uploaded:", verbose_level=2)
         self.output(template_contents, verbose_level=2)
 
-        # get existing scope if --retain-existing-scope is set
-        object_type = "os_x_configuration_profile"
-        if self.retain_scope and obj_id > 0:
-            self.output("Substituting existing scope into template", verbose_level=1)
-            existing_scope = self.get_existing_scope(
-                self.jamf_url, object_type, obj_id, token
-            )
-            # substitute pre-existing scope
-            template_contents = self.replace_scope(template_contents, existing_scope)
-
         self.output("Uploading Configuration Profile...")
         # write the template to temp file
         template_xml = self.write_temp_file(template_contents)
 
         # if we find an object ID we put, if not, we post
+        object_type = "configuration_profile"
         url = "{}/{}/id/{}".format(jamf_url, self.api_endpoints(object_type), obj_id)
 
         count = 0
@@ -272,48 +199,37 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
         return r
 
     def execute(self):
-        """Upload a configuration profile"""
+        """Upload a mobile device configuration profile"""
         self.jamf_url = self.env.get("JSS_URL")
         self.jamf_user = self.env.get("API_USERNAME")
         self.jamf_password = self.env.get("API_PASSWORD")
         self.client_id = self.env.get("CLIENT_ID")
         self.client_secret = self.env.get("CLIENT_SECRET")
         self.profile_name = self.env.get("profile_name")
-        self.payload = self.env.get("payload")
         self.mobileconfig = self.env.get("mobileconfig")
         self.identifier = self.env.get("identifier")
         self.template = self.env.get("profile_template")
         self.profile_category = self.env.get("profile_category")
         self.organization = self.env.get("organization")
         self.profile_description = self.env.get("profile_description")
-        self.profile_computergroup = self.env.get("profile_computergroup")
+        self.profile_mobiledevicegroup = self.env.get("profile_mobiledevicegroup")
         self.replace = self.env.get("replace_profile")
-        self.retain_scope = self.env.get("retain_scope")
         self.sleep = self.env.get("sleep")
         self.unsign = self.env.get("unsign_profile")
-        # handle setting replace in overrides
-        if not self.replace or self.replace == "False":
-            self.replace = False
-        # handle setting retain_scope in overrides
-        if not self.retain_scope or self.retain_scope == "False":
-            self.retain_scope = False
         # handle setting unsign in overrides
         if not self.unsign or self.unsign == "False":
             self.unsign = False
+        # handle setting replace in overrides
+        if not self.replace or self.replace == "False":
+            self.replace = False
 
         # clear any pre-existing summary result
-        if "jamfcomputerprofileuploader_summary_result" in self.env:
-            del self.env["jamfcomputerprofileuploader_summary_result"]
+        if "JamfMobileDeviceProfileUploader_summary_result" in self.env:
+            del self.env["JamfMobileDeviceProfileUploader_summary_result"]
 
         profile_updated = False
 
         # handle files with no path
-        if self.payload and "/" not in self.payload:
-            found_payload = self.get_path_to_file(self.payload)
-            if found_payload:
-                self.payload = found_payload
-            else:
-                raise ProcessorError(f"ERROR: Payload file {self.payload} not found")
         if self.mobileconfig and "/" not in self.mobileconfig:
             found_mobileconfig = self.get_path_to_file(self.mobileconfig)
             if found_mobileconfig:
@@ -373,26 +289,6 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
             except KeyError:
                 organization = ""
 
-        # otherwise we are dealing with a payload plist and we need a few other bits of info
-        else:
-            if not self.profile_name:
-                raise ProcessorError("ERROR: No profile name supplied - cannot import")
-            if not self.payload:
-                raise ProcessorError(
-                    "ERROR: No path to payload file supplied - cannot import"
-                )
-            if not self.identifier:
-                raise ProcessorError(
-                    "ERROR: No identifier for mobileconfig supplied - cannot import"
-                )
-            mobileconfig_name = self.profile_name
-            description = ""
-            organization = ""
-
-        # we provide a default template which has no category or scope
-        if not self.template:
-            self.template = "Jamf_Templates/ProfileTemplate-no-scope.xml"
-
         # automatically provide a description and organisation from the mobileconfig
         # if not provided in the options
         if not self.profile_description:
@@ -400,7 +296,8 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
                 self.profile_description = description
             else:
                 self.profile_description = (
-                    "Config profile created by AutoPkg and JamfComputerProfileUploader"
+                    "Config profile created by AutoPkg and "
+                    "JamfMobileDeviceProfileUploader"
                 )
         if not self.organization:
             if organization:
@@ -426,7 +323,7 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
         else:
             raise ProcessorError("ERROR: Credentials not supplied")
 
-        obj_type = "os_x_configuration_profile"
+        obj_type = "configuration_profile"
         obj_name = mobileconfig_name
         obj_id = self.get_api_obj_id_from_name(
             self.jamf_url,
@@ -444,7 +341,6 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
                     existing_uuid,
                     existing_identifier,
                 ) = self.get_existing_uuid_and_identifier(self.jamf_url, obj_id, token)
-
                 if self.mobileconfig:
                     # need to inject the existing payload identifier to prevent ghost profiles
                     mobileconfig_contents = (
@@ -454,15 +350,7 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
                     )
                     mobileconfig_plist = plistlib.dumps(mobileconfig_contents)
                 else:
-                    # generate the mobileconfig from the supplied payload
-                    mobileconfig_plist = self.make_mobileconfig_from_payload(
-                        self.payload,
-                        self.identifier,
-                        mobileconfig_name,
-                        self.organization,
-                        self.profile_description,
-                        existing_uuid,
-                    )
+                    self.output("A mobileconfig was not supplied.")
 
                 # now upload the mobileconfig by generating an XML template
                 if mobileconfig_plist:
@@ -472,16 +360,13 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
                         self.profile_description,
                         self.profile_category,
                         mobileconfig_plist,
-                        self.profile_computergroup,
+                        self.profile_mobiledevicegroup,
                         template_contents,
                         existing_uuid,
                         token,
-                        self.retain_scope,
                         obj_id=obj_id,
                     )
                     profile_updated = True
-                else:
-                    self.output("A mobileconfig was not generated so cannot upload.")
             else:
                 self.output(
                     "Not replacing existing Configuration Profile. "
@@ -493,17 +378,6 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
             )
             new_uuid = str(uuid.uuid4())
 
-            if not self.mobileconfig:
-                # generate the mobileconfig from the supplied payload
-                mobileconfig_plist = self.make_mobileconfig_from_payload(
-                    self.payload,
-                    self.identifier,
-                    mobileconfig_name,
-                    self.organization,
-                    self.profile_description,
-                    new_uuid,
-                )
-
             # now upload the mobileconfig by generating an XML template
             if mobileconfig_plist:
                 self.upload_mobileconfig(
@@ -512,10 +386,10 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
                     self.profile_description,
                     self.profile_category,
                     mobileconfig_plist,
-                    self.profile_computergroup,
+                    self.profile_mobiledevicegroup,
                     template_contents,
                     new_uuid,
-                    token,
+                    token=token,
                 )
                 profile_updated = True
             else:
@@ -527,7 +401,7 @@ class JamfComputerProfileUploaderBase(JamfUploaderBase):
         self.env["profile_name"] = self.profile_name
         self.env["profile_updated"] = profile_updated
         if profile_updated:
-            self.env["jamfcomputerprofileuploader_summary_result"] = {
+            self.env["JamfMobileDeviceProfileUploader_summary_result"] = {
                 "summary_text": (
                     "The following configuration profiles were uploaded to "
                     "or updated in Jamf Pro:"
