+++ JamfUploaderProcessors/JamfUploaderLib/JamfPackageUploaderBase.py
@@ -1,10 +1,22 @@
 #!/usr/local/autopkg/python
 
 """
-Base classes for uploading a package to Jamf Pro
-    by G Pugh
+Copyright 2023 Graham Pugh
 
-Note the requirements for uploading to the JCDS2 API endpoint:
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+NOTES:
+Requirements for uploading to the JCDS2 API endpoint:
 - boto3
 
 To resolve the dependencies, run: /usr/local/autopkg/python -m pip install boto3
@@ -14,7 +26,6 @@ import hashlib
 import json
 import os.path
 import shutil
-import subprocess
 import sys
 import threading
 
@@ -116,41 +127,6 @@ class JamfPackageUploaderBase(JamfUploaderBase):
         return zip_name
 
     """Beginning of section for upload to Local Fileshare Distribution Points"""
-
-    def mount_smb(self, mount_share, mount_user, mount_pass):
-        """Mount distribution point."""
-        mount_cmd = [
-            "/usr/bin/osascript",
-            "-e",
-            (
-                f'mount volume "{mount_share}" as user name "{mount_user}" '
-                f'with password "{mount_pass}"'
-            ),
-        ]
-        self.output(
-            f"Mount command: {' '.join(mount_cmd)}",
-            verbose_level=4,
-        )
-
-        r = subprocess.check_output(mount_cmd)
-        self.output(
-            r.decode("ascii"),
-            # r,
-            verbose_level=4,
-        )
-
-    def umount_smb(self, mount_share):
-        """Unmount distribution point."""
-        path = f"/Volumes{urlparse(mount_share).path}"
-        cmd = ["/usr/sbin/diskutil", "unmount", path]
-        try:
-            r = subprocess.check_output(cmd)
-            self.output(
-                r.decode("ascii"),
-                verbose_level=2,
-            )
-        except subprocess.CalledProcessError:
-            self.output("WARNING! Unmount failed.")
 
     def check_local_pkg(self, mount_share, pkg_name):
         """Check local DP or mounted share for existing package"""
