+++ JamfUploaderProcessors/README.md
@@ -1,12 +1,16 @@
 # JamfUploader processors for AutoPkg
 
-These processors are copied from [grahampugh/jamf-upload](https://github.com/grahampugh/jamf-upload/tree/main/JamfUploaderProcessors). 
-All development of the processors is done there, and changes are copied in here.
+These processors are developed in [grahampugh/jamf-upload][1] and periodicially copied into [autopkg/grahampugh-recipes][2].
 
-I welcome contributions to this project. **Please issue PRs to the above project rather than this one.**
+I welcome contributions to this project. **Please only issue PRs to [grahampugh/jamf-upload][1]**
 
 ## Documentation
 
-Details of these processors can be found in the [grahampugh/jamf-upload wiki](https://github.com/grahampugh/jamf-upload/wiki/JamfUploader-AutoPkg-Processors).
+Details of these processors can be found in the [wiki][3].
 
-There are also descriptions of the input and output variables in the [READMEs folder](READMEs/).
+There are also descriptions of the input and output variables in the [READMEs folder][4].
+
+[1]: https://github.com/grahampugh/jamf-upload/tree/main/JamfUploaderProcessors
+[2]: https://github.com/autopkg/grahampugh-recipes/tree/main/JamfUploaderProcessors
+[3]: https://github.com/grahampugh/jamf-upload/wiki/JamfUploader-AutoPkg-Processors
+[4]: READMEs/
